Revert "201-add-basic-flashcards"
