--- components/Home.js
@@ -1,14 +1,0 @@
-import {View, Text, StyleSheet} from "react-native";
-import colors from "../styles/colors";
-
-const Home = () => {
-    return <View><Text>Home</Text></View>
-};
-
-const styles = StyleSheet.create({
-    home: {
-        color: colors.textColor
-    }
-});
-
-export default Home;
